chore: rename and restructure auth middlewares
