+++ src/Deployment/Deployment.router.ts
@@ -3,9 +3,10 @@ import Ajv from 'ajv'
 
 import { Router } from '../common/Router'
 import { HTTPError } from '../common/HTTPError'
-import { authn, AuthRequest, authz } from '../middleware/auth'
+import { authentication, AuthRequest } from '../middleware/authentication'
 import { Deployment } from './Deployment.model'
 import { DeploymentAttributes, deploymentSchema } from './Deployment.types'
+import { projectAuthorization } from '../middleware/authorization/project'
 
 const ajv = new Ajv()
 
@@ -16,8 +17,7 @@ export class DeploymentRouter extends Router {
      */
     this.router.get(
       '/deployments',
-      authn,
-      authz,
+      authentication,
       server.handleRequest(this.getDeployments)
     )
 
@@ -26,8 +26,8 @@ export class DeploymentRouter extends Router {
      */
     this.router.get(
       '/projects/:id/deployment',
-      authn,
-      authz,
+      authentication,
+      projectAuthorization,
       server.handleRequest(this.getProjectDeployment)
     )
 
@@ -36,8 +36,8 @@ export class DeploymentRouter extends Router {
      */
     this.router.put(
       '/projects/:id/deployment',
-      authn,
-      authz,
+      authentication,
+      projectAuthorization,
       server.handleRequest(this.upsertDeployment)
     )
 
@@ -46,8 +46,8 @@ export class DeploymentRouter extends Router {
      */
     this.router.delete(
       '/projects/:id/deployment',
-      authn,
-      authz,
+      authentication,
+      projectAuthorization,
       server.handleRequest(this.deleteDeployment)
     )
   }
